make tests for stack
